requirements.txt
ipython
python-rtmidi
mido
adafruit-circuitpython-ble-midi